REFACTOR: 소스 코드 구조 변경
