+++ src/main/java/io/whatap/task/order/entity/Order.java
@@ -1,4 +1,4 @@
-package io.whatap.task.entity;
+package io.whatap.task.order.entity;
 
 import io.quarkus.hibernate.orm.panache.PanacheEntity;
 import jakarta.persistence.Column;
